feat: Add support for run checks in State class
The code changes introduce support for run checks in the State class. Run checks are conditions that are checked on each tick to determine if the state should transition to another state or if another action should be executed. This feature enhances the flexibility and control of state transitions in the state machine.

Refactor state transition queue and update default_input_processor in eila.ts